requirements.txt
PyAudio==0.2.13
vosk==0.3.45
keyboard==0.13.5
Flask==3.0.0
pywin32==310
pystray==0.19.5
pillow==11.3.0
python-dotenv==1.0.0
